requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SSLify==0.1.5
fuzzywuzzy==0.18.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Levenshtein==0.26.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
pillow==11.1.0
PyAutoGUI==0.9.54
pycryptodome==3.21.0
PyGetWindow==0.0.9
pymongo==4.11.1
PyMsgBox==1.0.9
pynput==1.7.8
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytweening==1.2.0
RapidFuzz==3.12.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3